Fix in-process DMP under gunicorn: rename server entrypoint app.py -> server.py
The Flask simulation server lived in app.py, so any WSGI launch that imports it
as the top-level module `app` (gunicorn app:app, waitress_app.py) shadowed the
dmp/app package. InProcessDMP() then raised "No module named 'app.state_machine'"
and was silently caught at infectionmgr.py, falling back to the per-infection
HTTP DMP path (~32k round-trips) -- the original bottleneck. `python app.py` was
unaffected (server runs as __main__), so local dev stayed fast (~30s) while a
gunicorn-launched prod regressed to ~2m10.

Rename the entrypoint to server.py so the top-level name `app` always resolves
to dmp/app under any launch mode (python, gunicorn, waitress). Also:
- Dockerfile: run `gunicorn server:app` (multi-process concurrency) with
  --timeout 0 for the long SSE streams; WEB_CONCURRENCY defaults to 4.
- requirements: add gunicorn==23.0.0.
- config: accept DMP_API_URL as an alias for DMP_API_BASE_URL so the HTTP
  fallback still reaches the real dmp service (defense in depth).
- waitress_app.py / README: import server instead of app.

Verified: InProcessDMP constructs and samples timelines under a gunicorn
server:app launch; the old app:app launch reproduces the failure.
