+++ requirements.txt
@@ -20,6 +20,7 @@ fuzzywuzzy==0.18.0
 geopandas==1.1.1
 gitdb==4.0.12
 GitPython==3.1.46
+gunicorn==23.0.0
 h11==0.16.0
 haversine==2.9.0
 idna==3.11
